Added README

+++ Source/NewVegas/Buildings.cpp
@@ -18,8 +18,8 @@ Buildings::Buildings()
 
     Streets *streets = Streets::GetInstance();
     Parks *parks = Parks::GetInstance();
-    for (size_t i = 1; i < streets->horizStreets.size(); i++)
-        for (size_t j = 1; j < streets->vertStreets.size(); j++)
+    for (unsigned int i = 1; i < streets->horizStreets.size(); i++)
+        for (unsigned int j = 1; j < streets->vertStreets.size(); j++)
             if (!parks->hasPark.count(j * 1000 + i))
                 buildings.push_back(GenerateBuilding(streets->ZoneBetweenStreets(j - 1, j, i - 1, i)));
 }
